Fix issues 172, 175, 176, 178, 179
#172 build warnings. The audited set now means the same thing on every
compiler: GCC's -Wconversion does not imply -Wsign-conversion but Clang's
does, so the identical flag string audited ~3200 more sites under
Clang/AppleClang/emscripten than under the GCC the gate runs on. Sign
conversion becomes its own opt-in switch (OPENGLAD_WARN_SIGN_CONVERSION,
off) and every remaining warning is fixed: 24 under GCC 15 and 227 more
unique sites under Clang 21, plus the CMP0205 dev warning, the emscripten
suppressions that were ordered ahead of -Wall and silently re-enabled by it,
the 150 em++ experimental-port notices, and -Wpedantic auditing fetched
dependency headers (FetchContent SYSTEM). Three targets that linked no
warning flags at all now do: the fuzz targets, scenario_facts_dump and
og_lua_lines. OPENGLAD_WARNINGS_AS_ERRORS is on in the ci-test, ci-asan and
ci-tsan presets so it stays fixed. Two real defects fell out: the editor's
rectangle-select tested fabs(h > 15) instead of fabs(h) > 15, so an upward
drag never committed, and scenario_facts_dump's unaudited switch was missing
a FactKind.

#175 tick-0 snapshot hash mismatch. GameServer::bind_player only claimed the
control when it auto-selected one, so an installer-supplied control went out
in the initial keyframe still reading user == -1 while the same keyframe
named it as the player's control. The first tick claimed it, which made tick
0 the one snapshot that never matched the mirror and cost the desync
detector its first tick. The claim now applies to a supplied control too,
re-checked against the same §4.4 ownership policy the scans apply and
skipped for a dead walker (a peer reconnecting after its released walker
died binds against a corpse).

#176 pack-family base HP. gloader read hitpoints[PIX(order, family)] instead
of the loader slot it had already computed; PIX cannot address a family id
>= NUM_FAMILIES and aliases onto an unrelated core entry. In-bounds, and
both slots read 0 for the only shipped family above the core span, so no
shipped content changes.

#178 generators run with hp > max_hp. walker::set_difficulty's generator arm
wrote the fighting hp and left max_hitpoints on the loader base (0 for
tower/bones/treehouse), so a damaged generator could never regenerate, the
HP bar had no fraction to draw, and hp = max_hp emptied the post. Both
fields move together now, and the modes campaign's Lua workaround is
removed. Deliberate divergence from classic, recorded in
GAMEPLAY_FIXES_FROM_CLASSIC.md.

#179 on_damage contract. Returning a number - including 0 - is a
replacement amount, not a cancel: the hit still lands and walker::attack's
own hp <= 0 check still runs, so 0 against a 0-hp target kills it. Only
false cancels. Written down in family_hooks.h, the modding API reference,
the design doc, the generated LuaLS stub and the modding skill, and pinned
by four new gate tests.

Zero parity goldens move. og_test_parity is 256/256 and
tests/parity/golden/ is untouched: the harness never calls set_difficulty,
which the companion confirms from its own side - all 11 generator-bearing
scenarios recapture byte-identically with and without the fix. The
parity-companion branch carries the same one-line change (a2714ec5) so the
reference stays comparable.

+++ src/platform/sdl/sai2x.cpp
@@ -926,8 +926,8 @@ void Screen::set_active_canvas(CanvasTarget target)
 	{
 		const auto [x, y] = window_to_active_canvas(
 			pointer_window.first, pointer_window.second);
-		mouse_state.x = static_cast<int>(x);
-		mouse_state.y = static_cast<int>(y);
+		mouse_state.x = static_cast<float>(static_cast<int>(x));
+		mouse_state.y = static_cast<float>(static_cast<int>(y));
 	}
 }
 
@@ -1086,8 +1086,8 @@ bool Screen::set_world_canvas_size(int w, int h)
 	{
 		const auto [x, y] = window_to_active_canvas(
 			pointer_window.first, pointer_window.second);
-		mouse_state.x = static_cast<int>(x);
-		mouse_state.y = static_cast<int>(y);
+		mouse_state.x = static_cast<float>(static_cast<int>(x));
+		mouse_state.y = static_cast<float>(static_cast<int>(y));
 	}
 	if (previous_texture != ui_tex_)
 		SDL_DestroyTexture(previous_texture);
